Move work to compiler (#6)
* Move work to compiler

Previously dicts were parsed and stored in ETS on application startup.
This commit introduces modules to get vendors, attributes and values.
This saves time during startup of the application. Secondly this opens
the possibility to offer a set of macros that will replace the macro
with the correct value at compile time instead of the encoding phase at
runtime. This should also decrease the amount of times you run into
EntryNotFoundError.

nimble_parsec is compiled into a module so that users of the library
don't have to add nimble_parsec as a dep themselves.

* Add option to specify extra dictionaries outside of this repo

* Only filter included dictionaries when the library is configured

* Fixes `example.exs`